+++ package.js
@@ -35,6 +35,7 @@ function configure( api ){
     api.use( 'pwix:field@1.0.0-rc' );
     api.use( 'pwix:forms@1.0.0-rc' );
     api.use( 'pwix:i18n@1.5.7' );
+    api.use( 'pwix:image-includer@1.0.0-rc' );
     api.use( 'pwix:modal@2.0.0' );
     api.use( 'pwix:notes@1.0.0-rc' );
     api.use( 'pwix:roles@1.3.0' );